antes de liada

+++ app/src/main/java/com/example/animecalendar/model/NotificationItem.java
@@ -5,9 +5,9 @@ import java.util.List;
 public class NotificationItem {
 
     private List<String> animeTitles;
-    private long notificationTime;
+    private int notificationTime;
 
-    public NotificationItem(List<String> animeTitles, long notificationTime) {
+    public NotificationItem(List<String> animeTitles, int notificationTime) {
         this.animeTitles = animeTitles;
         this.notificationTime = notificationTime;
     }
@@ -20,11 +20,11 @@ public class NotificationItem {
         this.animeTitles = animeTitles;
     }
 
-    public long getNotificationTime() {
+    public int getNotificationTime() {
         return notificationTime;
     }
 
-    public void setNotificationTime(long notificationTime) {
+    public void setNotificationTime(int notificationTime) {
         this.notificationTime = notificationTime;
     }
 }
